+++ src/kernels/naive.rs
@@ -10,7 +10,7 @@ impl GemmKernel for NaiveGemm {
         "naive-ijk"
     }
 
-    fn compute(&self, lhs: &Matrix<f64>, rhs: &Matrix<f64>, output: &mut Matrix<f64>) {
+    fn compute(&self, lhs: &Matrix<f32>, rhs: &Matrix<f32>, output: &mut Matrix<f32>) {
         assert_gemm_dimensions(lhs, rhs, output);
         let n = lhs.cols();
