+++ src/kernels/tiled.rs
@@ -21,7 +21,7 @@ impl GemmKernel for TiledGemm {
         "tiled"
     }
 
-    fn compute(&self, lhs: &Matrix<f64>, rhs: &Matrix<f64>, output: &mut Matrix<f64>) {
+    fn compute(&self, lhs: &Matrix<f32>, rhs: &Matrix<f32>, output: &mut Matrix<f32>) {
         assert_gemm_dimensions(lhs, rhs, output);
         let n = lhs.cols();
         output.as_mut_slice().fill(0.0);
